Lexi/requirements.txt
absl-py==2.3.1
accelerate==0.21.0
addict==2.4.0
aenum==3.1.16
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
asgiref==3.9.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blendmodes==2022
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
PyPika @ git+https://github.com/kayak/pypika@master
chromadb==0.5.0
clean-fid==0.1.35
click==8.2.1
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
cmake==4.0.3
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
deepdanbooru==1.0.3
deprecation==2.1.0
diffusers==0.30.1
diskcache==5.6.3
distro==1.9.0
durationpy==0.10
einops==0.4.1
exceptiongroup==1.3.0
ffmpy==0.6.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.32
glcontext==3.0.0
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==3.41.2
gradio_client==0.5.0
greenlet==3.2.3
h5py==3.14.0
httptools==0.6.4
huggingface==0.0.1
humanfriendly==10.0
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.6
jsonmerge==1.9.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
kornia==0.6.7
kubernetes==33.1.0
lark==1.1.2
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.14.3
lit==18.1.8
llvmlite==0.44.0
lmdb==1.7.2
lycoris==0.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==5.1.0
moderngl==5.12.0
mpmath==1.3.0
multidict==6.6.3
multipledispatch==1.0.0
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.45.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.3.1
onnxruntime==1.22.1
onnxruntime-gpu==1.20.1
open_clip_torch==3.0.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.2.3
peft==0.7.1
piexif==1.1.3
pillow-avif-plugin==1.4.3
platformdirs==4.3.8
playwright==1.54.0
posthog==6.1.0
propcache==0.3.2
psutil==5.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydub==0.25.1
pyee==13.0.0
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
pyre-extensions==0.0.29
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
pytorch-lightning==1.9.4
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
requests-oauthlib==2.0.0
resize-right==0.0.2
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
scikit-image==0.21.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spandrel==0.3.4
spandrel_extra_arches==0.1.1
sympy==1.14.0
tb-nightly==2.21.0a20250723
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tifffile==2025.5.10
timm==1.0.19
tomesd==0.1.3
toml==0.10.2
tomli==2.2.1
torchdiffeq==0.2.3
torchmetrics==1.7.3
torchsde==0.2.6
tqdm==4.67.1
trampoline==0.1.2
typer==0.16.0
typing-inspect==0.9.0
tzdata==2025.2
urllib3==2.5.0
uvloop==0.21.0
voluptuous==0.15.2
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
httpx[http2]==0.27.2
redis==5.0.* 
gunicorn==21.2.0
uvicorn[standard]==0.30.6
APScheduler==3.10.4
tzlocal==5.3.1
pydantic-settings>=2.2
